+++ packages/manager/src/components/EntityHeader/index.tsx
@@ -1,5 +1,5 @@
-export { default } from './EntityHeader';
-import { HeaderProps as _HeaderProps } from './EntityHeader';
-
-// eslint-disable-next-line
+import EntityHeader, { HeaderProps as _HeaderProps } from './EntityHeader';
+/* tslint:disable */
 export interface HeaderProps extends _HeaderProps {}
+
+export default EntityHeader;