+++ packages/manager/src/features/linodes/LinodesLanding/LinodesLanding.test.tsx
@@ -15,9 +15,13 @@ describe('ListLinodes', () => {
     CSVlink: '',
     addNewLink: '',
     chip: '',
+    chipActive: '',
     chipRunning: '',
     chipPending: '',
-    chipOffline: ''
+    chipOffline: '',
+    controlHeader: '',
+    toggleButton: '',
+    clearFilters: ''
   };
 
   it('renders without error', () => {
@@ -46,6 +50,7 @@ describe('ListLinodes', () => {
           {...reactRouterProps}
           ldClient={{} as any}
           flags={{}}
+          linodesInTransition={new Set<number>()}
         />
       )
     );